% Generated by roxygen2 (4.0.2): do not edit by hand
\name{applyDToD}
\alias{applyDToD}
\title{Apply a function from one calendar day to the next}
\usage{
applyDToD(x, fn, dayNum = 2, type = "week", parallel = FALSE, ...)
}
\arguments{
\item{x}{input xts}

\item{fn}{function to be applied}

\item{dayNum}{is the number that matches to the day wanted. For type "week",
the numbers go from 1 (sunday) to 7 (saturday). For type "month" they go from
1 to 31.}

\item{type}{week or month, where default is week. For example , from Monday
(dayNum=2) to Monday for "week" or from the 2rd to the next 2rd for "month".}

\item{parallel}{default is FALSE. If TRUE use snowfall package to run in
parallel.}

\item{...}{additional parameters for function fn}
}
\description{
One can apply a function from one day of the week to the next. So for
example, from one Monday to the next Monday. Or from the 3rd of the month to
the next 3rd.  Notice that if your data does not have a 3rd every month (for
instance it is a holiday), then apply will skip to the next 3rd available.
This results into a gap of at least 2 months. The same goes for weekly data.
}
\examples{
x1 <- xts(1:100,Sys.Date()+1:100)
x2 <- merge(x1,xts(101:200,Sys.Date()+1:100))
# Using multiple CPUs with snowfal
\dontrun{
library(snowfall)
sfInit(parallel=TRUE,cpus=4)
sfLibrary(xts)
applyDToD(x1,fn=mean,dayNum=2,parallel=TRUE)
applyDToD(x2,fn=colMeans,dayNum=2,parallel=TRUE)
sfStop()
}
# Not parallel
applyDToD(x1,fn=mean,dayNum=3)
applyDToD(x2,fn=colMeans,dayNum=3)
}

