% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bDays}
\alias{bDays}
\title{Buissines calendar using RQuantLib}
\usage{
bDays(from = "1980-01-01", to = Sys.Date(), by = 1,
  calendar = "UnitedStates/NYSE", useYahoo = FALSE)
}
\arguments{
\item{from}{starting day}

\item{to}{end of period}

\item{by}{step to update days}

\item{calendar}{name of calendar. See ?isBusinessDay in RQuantLib for
possible names. Default is "UnitedStates/NYSE".}

\item{useYahoo}{do you want to download the NYSE calendar from Yahoo? It
works only from days after 1950. Default is FALSE.}
}
\value{
array of buisness days
}
\description{
Using QuantLib calendars as default. However we found some errors comparing
to NYSE. Therefore we have an option to use the S&P500 index from yahoo to
extract the valid buissines days for NYSE calendar. Yahoo has data
after 1950. For data before 1950 we use QuantLib. QuantLib NYSE calendar after
1980 is corrected manually on 2014-09. That means that Yahoo and NYSE dates
are going to match, but since it is manual process (I give it a list of days)
it might not be updated fast enough, therefore there could be errors!
Notice that you
will need internet to download data from yahoo. It
is also slower to download data from yahoo. Finally, Yahoo data might also
have errors (missing business days or other)...
According to RQuantLib documentation there are lots of other calendars. See
?isBusinessDay for other calendars (bond, energy and other countries)
}
\examples{
b <- bDays()
}

