% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{calculate}
\alias{calculate}
\title{Method to calculate the current value of a rolltable}
\usage{
calculate(obj, ...)
}
\arguments{
\item{obj}{Table to calculate, such as a rolltable.}

\item{...}{For use by class methods}
}
\value{
A numeric vector whose names are the individual sums for each dice set.
}
\description{
Method to calculate the current value of a rolltable
}
\examples{

tbl <- rolltable("2d20h1")
calculate(tbl)
calculate(tbl, .summary_fn = "median")
}
