% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{calculate.rolltable}
\alias{calculate.rolltable}
\title{Calculate a rolltable}
\usage{
\method{calculate}{rolltable}(obj, .summary_fn = "identity", ...)
}
\arguments{
\item{obj}{A rolltable.}

\item{.summary_fn}{A character string providing the name of a summary function, such as mean or median, or the function itself.}

\item{...}{Unused.}
}
\value{
A rolltable_calculation.
}
\description{
Summarizes the provided summary function over repetitions, groups by Die if more than one in the table.
}
