% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_roll_cmd.R
\name{evaluate_roll_cmd}
\alias{evaluate_roll_cmd}
\title{A complex roll resolution}
\usage{
evaluate_roll_cmd(parsed_cmd)
}
\arguments{
\item{parsed_cmd}{a list of dice and operators as returned by parse_roll_cmd.}
}
\value{
roll result.
}
\description{
@description The complex roll resolution function evaluates each elements of a complex
 roll command and compute their total result
}
