% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_dice.R
\name{infinite_letters}
\alias{infinite_letters}
\title{Helper function to create names using only letters, not numbers
So that a command can be substituted without accidentally overwriting the number.
For example, subbing "1d6 + 1" with tbls[[1]] and tbls[[2]] needs to result in tbls[[1]] + tbls[[2]]
But instead will result in tbls[[tbls[[2]]]] + 1}
\usage{
infinite_letters(n)
}
\arguments{
\item{n}{Number of distinct names required}
}
\value{
a character vector of unique names
}
\description{
Helper function to create names using only letters, not numbers
So that a command can be substituted without accidentally overwriting the number.
For example, subbing "1d6 + 1" with tbls[[1]] and tbls[[2]] needs to result in tbls[[1]] + tbls[[2]]
But instead will result in tbls[[tbls[[2]]]] + 1
}
