% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ma}
\alias{ma}
\title{Simple Moving average faster version}
\usage{
ma(x, n)
}
\arguments{
\item{x}{matrix or xts}

\item{n}{lookback window to average over}
}
\value{
matrix with MA calculated.
        The first n values are set to NA.
}
\description{
Uses RcppRoll function to implement MA.
Basically MA(2) = x(1)+x(2)/2; MA(3) = x(3)+x(2)/2 for window = 2
}
\examples{
ma(xts(1:10,Sys.Date()+1:10),n=2)
}

