% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{mean.rolltable}
\alias{mean.rolltable}
\title{Mean rolltable}
\usage{
\method{mean}{rolltable}(x, ...)
}
\arguments{
\item{x}{A rolltable class.}

\item{...}{Unused.}
}
\value{
A numeric vector whose names are the individual means for each dice set.
}
\description{
Determines the average over repetitions for the rolls.
Calculates the total by applying the operators to each average.
}
