% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{min.rolltable}
\alias{min.rolltable}
\title{Min rolltable}
\usage{
\method{min}{rolltable}(..., na.rm = FALSE)
}
\arguments{
\item{...}{A rolltable class.}

\item{na.rm}{Unused. For consistency with \code{\link[base]{min}}.}
}
\value{
A numeric vector whose names are the individual minimums for each dice set.
}
\description{
Determines the minimum over repetitions for the rolls.
Calculates the total by applying the operators to each minimum.
}
