% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{new_rolltable}
\alias{new_rolltable}
\title{Constructor for rolltable}
\usage{
new_rolltable(cmd, tbl)
}
\arguments{
\item{cmd}{Character vetor. Intended to be one or more dice set commands.}

\item{tbl}{Dataframe to convert to rolltable.}
}
\value{
Rolltable class object.
}
\description{
Internal function to build the rolltable class structure.
Minimal checks for validity.
}
