% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_calculation_class.R
\name{new_rolltable_calculation}
\alias{new_rolltable_calculation}
\title{Rolltable calculation class}
\usage{
new_rolltable_calculation(lst, n = names(lst))
}
\arguments{
\item{lst}{A list object returned using the "by" function.}

\item{n}{Names to use for the returned object, if the lst is not already named.}
}
\value{
A rolltable_calculation classed object.
}
\description{
Class returned when calculating a rolltable.
This inherits from the "by" class returned when using the \code{\link[base]{by}} function.
Typically, this function is not needed by the user, who instead should rely on \code{\link{calculate}}.
}
