% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_calculation_class.R
\name{Ops.rollr}
\alias{Ops.rollr}
\title{S3 Group Generic Ops for rolltable_calculation and rolltable}
\usage{
\method{Ops}{rollr}(e1, e2 = NULL)
}
\arguments{
\item{e1}{objects.}

\item{e2}{objects.}
}
\description{
Permits mathemetical operations on rolltables or rolltable calculations in certain circumstances.
Namely, the number of repetitions must match between objects, or one object is a length-one vector.
}
\details{
In all valid instances, the rolltable or rolltables are first calculated.
If there are multiple repetitions, these are all kept separate.
A rolltable can only be added, etc. to another rolltable with a comparable number of repetitions.
Repetitions will be increased to match, which will cause the rolltable to be rerolled.
If dies differ within a rolltable, each will be treated separately.
}
