% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{parse_result}
\alias{parse_result}
\title{Helper function for calculate to apply a summary function to a rolltable.}
\usage{
parse_result(tbl, .summary_fn = identity)
}
\arguments{
\item{tbl}{Rolltable}

\item{.summary_fn}{A summary function to apply}
}
\value{
A summarized result generated using the by function.
}
\description{
Helper function for calculate to apply a summary function to a rolltable.
}
