% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{print.rolltable}
\alias{print.rolltable}
\title{Print a rolltable}
\usage{
\method{print}{rolltable}(x, n = 10, rolls = FALSE, digits = 1, ...)
}
\arguments{
\item{x}{A rolltable class.}

\item{n}{Number of repetitions to print.}

\item{rolls}{If TRUE, print the base rolls and modifications to rolls, such as keep highest or exploding.}

\item{digits}{Round numbers to this number of digits when printing.}

\item{...}{Unused.}
}
\description{
Print a rolltable
}
