% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{repetitions}
\alias{repetitions}
\title{Number of repetitions in a rolltable}
\usage{
repetitions(tbl, ...)
}
\arguments{
\item{tbl}{A rolltable class.}
}
\value{
Integer value indicating the number of repetitions.
}
\description{
Number of repetitions in a rolltable
}
