% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resamplePeriodically}
\alias{resamplePeriodically}
\title{Resample a xts object into a matrix every N time periods.}
\usage{
resamplePeriodically(R, period = 5, na.rm = FALSE)
}
\arguments{
\item{R}{input xts time series.}

\item{period}{the number of periods to split the data. Default is 5.}

\item{na.rm}{if true we backfill with previus values and rm any NA that comes
from merging the xts time series together. Default = FALSE}
}
\value{
xts matrix with resampled time series
}
\description{
We can take a time series and separate the time series into observation every
N periods. For example if we have daily data and want to have data every 5
days, we can use this function.  This function will return (in the case of 5
periods)  xts matrix of 5 cols, one for data from 1 and 5, other from 2 to 6
and so forth. If one wants to reseample weekly or monthly one can use
xts build in functions such as to.weekly(x,OHLC=F), or apply.weely(x,last).
}
\examples{
resamplePeriodically(xts(1:100,Sys.Date()+1:100))
}

