% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{roll}
\alias{roll}
\title{Method to roll one or more iterations of dice.}
\usage{
roll(tbl, ...)
}
\arguments{
\item{tbl}{A rolltable class.}

\item{repetitions}{Number of times to roll the command.}

\item{verbose}{If TRUE, dice roll details are visible in the console.}
}
\value{
An updated rolltable.
}
\description{
Takes a rolltable and rolls to create a new set of results.
}
