% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{roll.rolltable}
\alias{roll.rolltable}
\title{Roll or reroll a rolltable}
\usage{
\method{roll}{rolltable}(obj, repetitions = 1, verbose = FALSE, ...)
}
\arguments{
\item{obj}{A rolltable.}

\item{repetitions}{Number of times to roll the command.}

\item{verbose}{If TRUE, dice roll details are visible in the console.}

\item{...}{Unused.}
}
\description{
Any existing repetitions will be removed and all rolls re-done.
}
\examples{

tbl <- rolltable("2d20h1"); tbl
roll(tbl, repetitions = 3)
roll(tbl, repetitions = 2, verbose = TRUE)

}
