% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_one.R
\name{roll_one}
\alias{roll_one}
\title{Roll a dice command}
\usage{
roll_one(dice, operator)
}
\arguments{
\item{dice}{a string corresponding to a dice command}
}
\value{
result of the roll(s)
}
\description{
Roll a dice command
}
