% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{rolltable}
\alias{rolltable}
\title{Rolltable class}
\usage{
rolltable(cmd, repetitions = 1, verbose = FALSE)
}
\arguments{
\item{cmd}{character string describing the dice roll to compute.
Can be of any length, but each character element can only represent a single dice set with no operators.}

\item{repetitions}{Number of times to roll the command.}

\item{verbose}{If TRUE, dice rolls details are visible in the console.}
}
\description{
Class used to store multiple dice rolls for a roll command.
This function is the main helper, and sets up a new rolltable.
}
\details{
Users interested in simulation may wish to use the rolltable class directly.
For example, the following simulates 1000 different D&D 5e ability rolls:
\code{rolltable("4d6h3", repetitions = 1000)}

You can apply mathematical operations to one or more rolltables. For example:
\code{(rolltable("3d6") * 3) + rolltable("1d4r1")}
\code{median(rolltable("4d6h3", repetitions = 1000))}
}
\examples{
rolltable("2d20h1")
rolltable("2d20h1", repetitions = 5) + rolltable("10")

# How often would advantage beat disadvantage in D&D 5e?
result <- rolltable("2d20h1", repetitions = 1000) > rolltable("2d20l1", repetitions = 1000)
summary(result[[1]])

# convert to dataframe
as.data.frame(rolltable("2d20h1", repetitions = 10))
as.data.frame(rolltable(c("1d6", "2d6", "3d6", "4d6"), repetitions = 2))

}
