% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.r
\name{sampleEOM}
\alias{sampleEOM}
\title{Resample a xts object into at the end of the month(EOM)}
\usage{
sampleEOM(R)
}
\arguments{
\item{R}{input xts time series.}
}
\value{
xts matrix with resampled time series
}
\description{
This function should have basically the same functionality to xts to.monthly with OHLC=FALSE
We have found occasional errors with xts::to.monthly most probably due to the date time class.
Therefore this code uses lubridate instead and works only for data without time info. We include the last
point of the time series even if it is not the end of the month...
}
\examples{
sampleEOM(xts(1:100,Sys.Date()+1:100))
}

