% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{trim_numeric_string}
\alias{trim_numeric_string}
\title{Helper function to change the number of digits for a string with numerals}
\usage{
trim_numeric_string(str, digits = 2)
}
\description{
Helper function to change the number of digits for a string with numerals
}
