% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{calculate}
\alias{calculate}
\title{Method to calculate the current value of a rolltable}
\usage{
calculate(tbl, ...)
}
\arguments{
\item{tbl}{A rolltable class.}

\item{.summary_fn}{Function used to summarize results between repetitions. Default returns each repetition separately.}
}
\value{
A numeric vector whose names are the individual sums for each dice set.
}
\description{
For each repetition in the rolltable, summarizes each dice set and
applies the operators in turn.
}
\examples{

tbl <- rolltable("2d20h1 + 20")
calculate(tbl)
calculate(tbl, .summary_fn = "median")
}
