% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{mean.rolltable}
\alias{mean.rolltable}
\title{Determines the average over repetitions for the rolls.
Calculates the total by applying the operators to each average.}
\usage{
\method{mean}{rolltable}(x)
}
\arguments{
\item{tbl}{A rolltable class.}
}
\value{
A numeric vector whose names are the individual means for each dice set.
}
\description{
Determines the average over repetitions for the rolls.
Calculates the total by applying the operators to each average.
}
