% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{median.rolltable}
\alias{median.rolltable}
\title{Median rolltable}
\usage{
\method{median}{rolltable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{A rolltable class.}
}
\value{
A numeric vector whose names are the individual means for each dice set.
}
\description{
Determines the median over repetitions for the rolls.
Calculates the total by applying the operators to each median value.
}
