% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{new_rolltable}
\alias{new_rolltable}
\title{Constructor for rolltable}
\usage{
new_rolltable(lst, cmd, dices, operators)
}
\description{
Internal function to build the rolltable class structure.
Minimal checks for validity.
}
