% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{roll}
\alias{roll}
\title{Roll Method}
\usage{
roll(tbl, ...)
}
\arguments{
\item{tbl}{A rolltable class.}

\item{repetitions}{Number of times to roll the command.}

\item{verbose}{If TRUE, dice roll details are visible in the console.}
}
\value{
An updated rolltable.
}
\description{
Takes a rolltable and rolls to create a new set of results, using one or more repetitions.
}
\examples{

tbl <- rolltable("2d20h1 + 20"); tbl
roll(tbl, repetitions = 3)
roll(tbl, repetitions = 2, verbose = TRUE)
}
