% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{roll}
\alias{roll}
\title{Roll Method}
\usage{
roll(obj, ...)
}
\arguments{
\item{obj}{Object to roll}

\item{...}{For use by class methods}
}
\value{
An updated rolltable.
}
\description{
Method for rolling dice.
}
