% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_dice.R
\name{roll_dice}
\alias{roll_dice}
\title{Roll several dice and perform computations}
\usage{
roll_dice(cmd, roll_history = FALSE)
}
\arguments{
\item{cmd}{the string describing the dice roll to simulate}
}
\value{
result of the dice roll computation
}
\description{
Roll several dice and perform computations
}
\examples{

}
