% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_dice.R
\name{roll_dice}
\alias{roll_dice}
\title{Roll Dice}
\usage{
roll_dice(cmd, roll_history = FALSE, repetitions = 1, verbose = FALSE)
}
\arguments{
\item{cmd}{character string describing the dice roll to compute.}

\item{roll_history}{boolean, if TRUE, dice rolls details are visible in the console.}
}
\value{
result of the dice roll computation
}
\description{
Main function to use the library. This function wraps all the computing functions of the library. let the user choose between print or hide all rolls details in the console.
}
\examples{

roll_dice("1d10 + 20")
roll_dice("1d4 * 2")
roll_dice("2d20h1")

}
