% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_one.R
\name{roll_one}
\alias{roll_one}
\title{Roll One}
\usage{
roll_one(roll)
}
\arguments{
\item{roll}{a string corresponding to a roll command.}
}
\value{
result of the roll
}
\description{
Roll one die from string command.
}
