% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolltable_class.R
\name{rolltable}
\alias{rolltable}
\title{Rolltable class}
\usage{
rolltable(cmd, repetitions = 1, verbose = FALSE)
}
\arguments{
\item{cmd}{character string describing the dice roll to compute.}

\item{repetitions}{Number of times to roll the command.}

\item{verbose}{If TRUE, dice rolls details are visible in the console.}
}
\description{
Class used to store multiple dice rolls for a roll command.
This function is the main helper, which sets up a new rolltable.
}
